add wget require explicit dependency
We use wget on fedora-upgrade and it isn't installed
by default on fedora.

Bellow is the error message user would get
when running fedora-upgrade without having wget
on the system:
---
/sbin/fedora-upgrade: line 93: wget: command not found
argument is not an RPM package
cpio: premature end of file
---

This commit fix this issue, by explicit requiring wget
on RPM Spec file.