Update remove-retired-packages
DESCRIPTION

remove-retired-packages is a nice program but it does not have any code to check the arguments that it was called with.  By checking to see if remove-retired-packages was called without valid arguments, a useful message can be printed instead of an incidental error message.

STEPS TO REPRODUCE

run:  remove-retired-packages --help

or run: remove-retired-packages thirty-five

etc.

EXPECTED BEHAVIOR

I expect remove-retired-packages would display a useful message.

ACTUAL BEHAVIOR

remove-retired-packages displays an error message that is not helpful.