Speed up selecting retired installed packages
The main speed up comes from also creating a list of installed packages
and joining it against the list of retired packages (with `comm`),
because this eliminates rpm calls in two loop bodies.

With LC_ALL=C, the string comparison heavy sort/comm calls end up
calling memcmp() instead of much slower locale-aware code.

The first measure easily saves a few minutes even on a cloud
instance or fast laptop while the second one yields a speed up of 1.14 or so
on top. Concrete numbers depend on the number of retired packages
and the total number of (installed) packages, of course.
For example, tests between Fedora 35/37 and 36/38 show such speed ups.